feat: 메모 업데이트 엔터키 기능 완료, 파일 저장 경로 제작 중 style: delete 버튼 커서 변경 완료,

+++ src/renderer/components/body/BodyComponent.tsx
@@ -2,6 +2,7 @@
 import styled from 'styled-components';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
+import { confirm } from 'react-confirm-box';
 import DateStartToEnd from './DateStartToEnd';
 import ListTitleComponent from './ListTitleComponent';
 import ListContentComponent from './ListContentComponent';
@@ -161,16 +162,19 @@ const BodyComponent = () => {
     // arr.map((c, index) => console.log(`${index}@@@@allcheckedItem:${c}`));
   };
 
-  const onDeleteHandler = () => {
-    console.log('checkedItems', checkedItems);
-    const ids = [...checkedItems];
-    // 여기 DB 삭제
-    window.ipcDbChannel.deleteCallQureyToMain(ids, (result: any) => {
-      console.log('Call delete result:', result);
-      checkedItems.clear();
-      setCheckedItems(new Set());
-      setDeleteState(!deleteState);
-    });
+  const onDeleteHandler = async () => {
+    // if (window.confirm('Are you sure you want to delete?')) {
+    if (await confirm('Are you sure you want to delete?')) {
+      console.log('checkedItems', checkedItems);
+      const ids = [...checkedItems];
+      // 여기 DB 삭제
+      window.ipcDbChannel.deleteCallQureyToMain(ids, (result: any) => {
+        console.log('Call delete result:', result);
+        checkedItems.clear();
+        setCheckedItems(new Set());
+        setDeleteState(!deleteState);
+      });
+    }
   };
 
   console.log(`^^^^Rendering BodyComponent..${recordState}`);
